+++ src/context/index.ts
@@ -1,11 +1,2 @@
-export { default as Provider } from "./provider";
 export { default as Store } from "./store";
-
-export type {
-  SessionState,
-  SessionAction,
-  SessionStore,
-  SessionApi,
-} from "./session";
-
 export { SessionContext } from "./store";
