+++ src/hooks/useSession.ts
@@ -1,7 +1,8 @@
-import { type SessionStore, SessionContext } from "~/context";
+import type { SessionStore } from "~/lib/types";
 import { useContext } from "react";
 import { useStore } from "zustand";
 import { useShallow } from "zustand/shallow";
+import { SessionContext } from "~/context";
 
 export default function useSession<T>(selector: (store: SessionStore) => T) {
   const context = useContext(SessionContext);
